refactor(jsdoc): countries
