--- libs/@guardian/libs/src/countries/@types/Country.ts
@@ -1,3 +1,0 @@
-import type { countries } from '../countries';
-
-export type Country = typeof countries[keyof typeof countries];
