--- libs/@guardian/libs/src/countries/@types/CountryCode.ts
@@ -1,4 +1,0 @@
-import type { Country } from './Country';
-
-/** https://en.wikipedia.org/wiki/ISO_3166-1 */
-export type CountryCode = Country['countryCode'];
