--- libs/@guardian/libs/src/countries/@types/CountryKey.ts
@@ -1,3 +1,0 @@
-import type { countries } from '../countries';
-
-export type CountryKey = keyof typeof countries;
